feat: ansible-playbook.yml file
This Playbook file is used to automate building docker image, since we are using a lot of commands to build, remove, and push images also to stop, remove and run containers so this file contains all the commands that you are going to use while testing building, pushing an image and run container.

 The file should do the following:

1. Ensure that docker is is Started.

2. Stop & Remove the current running containers if exist

3. Build & Run the containers exist inside the docker-compose.yml file

So as you see this file is idempotent file that can run miltiple times without changing the output plus you can use only one image name and one container name so you don't get lost by the number of images and names that you are going to build with different names.

 Before run nany command on windows you must have wsl, and type wsl before any command you want to execute.